Remove Cinnamon layout selection
This was introduced a long time ago to offer a shortcut to revert
changes which were introduced when GWL was brought in.

It's been years since and this option makes no sense for new users.

All the settings in this can be configured individually and these
preferences are preserved on upgrades.